Test de application de Category Terminados

+++ symfony/src/Category/Application/UseCase/DeleteCategory/DeleteCategoryHandler.php
@@ -12,11 +12,13 @@ class DeleteCategoryHandler
         private CategoryRepositoryInterface $categoryRepository
     ) {}
 
-    public function __invoke(string $id): Category
+    public function __invoke(string $id): ?Category
     {
         $category = $this->categoryRepository->findById(new CategoryId($id));
         //Aqui se debería comprobar si es el id del usuario correcto con un servicio, en versiones de la aplicación cuando se integre el JWT
-        $this->categoryRepository->delete($category);
+        if ($category) {
+            $this->categoryRepository->delete($category);
+        }
         return $category;
     }
 }
